Move is_indexable properties from models to models/mixins
