+++ backend/apps/github/models/mixins/repository.py
@@ -11,6 +11,16 @@ class RepositoryIndexMixin:
     """Repository index mixin."""
 
     @property
+    def is_indexable(self):
+        """Repositories to index."""
+        return (
+            not self.is_archived
+            and not self.is_empty
+            and not self.is_template
+            and self.project_set.exists()
+        )
+
+    @property
     def idx_commits_count(self):
         """Return commits count for indexing."""
         return self.commits_count
