+++ backend/apps/github/models/release.py
@@ -50,11 +50,6 @@ class Release(BulkSaveModel, NodeModel, ReleaseIndexMixin, TimestampedModel):
         """Return release summary."""
         return f"{self.tag_name} on {self.published_at.strftime('%b %d, %Y')}"
 
-    @property
-    def is_indexable(self):
-        """Releases to index."""
-        return not self.is_draft
-
     def from_github(self, gh_release, author=None, repository=None):
         """Update instance based on GitHub release data."""
         field_mapping = {
