+++ backend/apps/github/models/user.py
@@ -5,7 +5,6 @@ from django.db import models
 from apps.common.models import TimestampedModel
 from apps.github.models.common import GenericUserModel, NodeModel
 from apps.github.models.mixins.user import UserIndexMixin
-from apps.github.models.organization import Organization
 
 
 class User(NodeModel, GenericUserModel, TimestampedModel, UserIndexMixin):
@@ -22,14 +21,6 @@ class User(NodeModel, GenericUserModel, TimestampedModel, UserIndexMixin):
     def __str__(self):
         """User human readable representation."""
         return f"{self.name or self.login}"
-
-    @property
-    def is_indexable(self):
-        """Users to index."""
-        return (
-            self.login != "ghost"  # See https://github.com/ghost for more info.
-            and self.login not in Organization.get_logins()
-        )
 
     @property
     def issues(self):
